cmake: Allow out-of-tree builds (#201)
* cmake: Allow out-of-tree builds

This is a best practice, allowing easier development (not having to worry about cleaning the source directory).

* cmake: Install desktop file to proper location

* cmake: Install locales to correct platform paths on Linux

We still need to keep MyApp::AddTranslation for cases when the app
is built with something other than CMake, and on other platforms than Linux.

We are also adding a TREESHEETS_RELOCATABLE_INSTALLATION CMake configuration flag
that allows installing to the old relocatable location even on Linux.

Finally, we only install the necessary `*.mo` files, not the `*.po` files or `*.bat` helper scripts.

* cmake: Add support for FHS paths for docs and data

This is needed to make Linux distributions happy.

Had to bump minimum macOS version to Catalina since std::filesystem::path
is not supported on older versions than 10.15. macOS 10.14 (Mojave)
is unsupported since October 2021.