Rescale images in image dropdown in case DPI changed (#377)
In order to refill the vector with the correctly scaled bitmaps when the DPI
changes, isolate the corresponding code block into a generic member function
called FillBitmapVector. This commit also places the variable imagepath into
the MyFrame class so that it is accessible for the function that handles
the DPI change event.