Reset layout when a single line of text is dropped (#569)
When a single line of text is dropped onto a cell, the layout is not reset.

This is not optimal when the text length is longer than e.g. the column width
because the text gets cut as the layout is not adapted.

This commit fixes this by resetting the layout in the case of a drop of
a single line of text.