Place Drag and Drop data structures per document
When a page gets closed, the destructor of DropTarget is called
that also frees the associated data. If we maintain only one data
structure associated with each DropTarget of each page, it tries to
free this data structure multiple times, resulting in a crash.

To avoid this, keep a Drag and Drop data structure in each document.